Place policy in own file
